feat: getAllPositiveBalance for wallet storage
