+++ src/screens/HomeStack/TransactionStack/transaction-finish.tsx
@@ -31,7 +31,6 @@ export const TransactionFinishScreen = observer(() => {
     TransactionStackParamList,
     TransactionStackRoutes.TransactionFinish
   >().params;
-  Logger.log('TransactionFinishScreen', {hash, transaction, token, amount});
   const contact = Contact.getById(transaction?.to ?? '');
 
   const short = useMemo(
