+++ src/screens/SwapStack/swap-screen.tsx
@@ -953,9 +953,12 @@ export const SwapScreen = observer(() => {
   };
 
   const onPressChangeWallet = useCallback(async () => {
+    const wallets = Wallet.getAll();
+    const walletsWithBalances = Wallet.getAllPositiveBalance();
+
     const address = await awaitForWallet({
       title: I18N.selectAccount,
-      wallets: Wallet.getAllVisible(),
+      wallets: walletsWithBalances?.length ? walletsWithBalances : wallets,
       initialAddress: currentWallet.address,
     });
     await refreshTokenBalances(address as HaqqEthereumAddress, tokenIn);
